# storage/api-docs/api-docs-admin-v1.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation for admin",
        "description": "L5 Swagger OpenApi description for admin",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/admin/v1/get-roles": {
            "get": {
                "tags": [
                    "Role"
                ],
                "summary": "Retrieve all roles",
                "operationId": "getAllRoles",
                "responses": {
                    "200": {
                        "description": "User updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Status indicating the success of the operation",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "Updated user data",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Status indicating the failure of the operation",
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKey": []
                    }
                ]
            }
        },
        "/admin/v1/user/{user_email}/permission/{permission}": {
            "patch": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Give permission to a user",
                "operationId": "givePermission",
                "parameters": [
                    {
                        "name": "user_email",
                        "in": "path",
                        "description": "User email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "permission",
                        "in": "path",
                        "description": "Permission ",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "object"
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKey": []
                    }
                ]
            }
        },
...
        "/admin/v1/show/{email}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Retrieve specific user",
                "operationId": "getUserByEmail",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "object"
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "APIKey": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "APIKey": {
                "type": "apiKey",
                "name": "X-Authorization",
                "in": "header"
            }
        }
    }
}